light: Use soong config declaration based on platform version
Since vendor is fixed on Android 11 and system varies, the shared
library dependency has to be split.

The ndk_platform backend will soon be deprecated because the ndk backend
can serve the same purpose. This is to eliminate the confusion about
having two variants (ndk and ndk_platform) for the same ndk backend.

So, for Android 13 and beyond, use android.hardware.light-V1-ndk
where as use android.hardware.light-ndk_platform for Android 11.

Change-Id: I255ac747f8a122fc90945ec21f584ae3ba364c68