Add support to build ABL as a part of build
While the ABL (Android BootLoader) is secure-boot signed and therefore
cannot be modified by non-Fairphone parties, allow building the image
from the tree, mostly for internal build purposes.

$ m out/target/product/FP4/abl.elf
$ sectools.py secimage --chipset bitra --sign_id abl -i out/target/product/FP4/abl.elf -s

Note for the public: Do not flash the image to your abl partitions, you
will not be able to recover the device.

Issue: FP4-A14#497
Change-Id: Ic89674e1d341b7ed240db8ccb64d550dbfb7737d